枚举治理，统一改为强类型枚举，删除无效的枚举定义

+++ source/msgui/windows/msgdlg/msgdlg.cpp
@@ -96,7 +96,7 @@ void WechatMsgDialog::initUI()
 
 void WechatMsgDialog::turnToPage(WechatPage page)
 {
-	m_verticalNavigationBar->setCurrentPage(page);
+	m_verticalNavigationBar->setCurrentPage(static_cast<unsigned int>(page));
 }
 
 STDMETHODIMP_(bool) WechatMsgDialog::OnCoreEvent(IJCoreEvent* event)
